Refactor the CLI for better readability and navigation
---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/and-other-tales/homework-mini?shareId=XXXX-XXXX-XXXX-XXXX).
