+++ cli/__init__.py
@@ -1,0 +1,1 @@
+# This is an empty __init__.py file to mark the cli directory as a package.
